+++ MiniProject/Form_Students.cs
@@ -139,15 +139,21 @@ namespace MiniProject
             if (senderGrid.Columns[e.ColumnIndex] is DataGridViewButtonColumn &&
                 e.RowIndex >= 0 && e.ColumnIndex == 7)
             {
-                int row_index = e.RowIndex;
-                DataGridViewRow selectedRow = dataGridView_Students.Rows[row_index];
-                string a = Convert.ToString(selectedRow.Cells["Id"].Value);
-                conn.Open();
-                string query = "DELETE FROM Student WHERE Id = '" + a + "'";
-                SqlCommand command = new SqlCommand(query, conn);
-                command.ExecuteNonQuery();
-                conn.Close();
-                loadDataGridView();
+                var confirmResult = MessageBox.Show("Are you sure to delete this item??",
+                                     "Confirm Delete!!",
+                                     MessageBoxButtons.YesNo);
+                if (confirmResult == DialogResult.Yes)
+                {
+                    int row_index = e.RowIndex;
+                    DataGridViewRow selectedRow = dataGridView_Students.Rows[row_index];
+                    string a = Convert.ToString(selectedRow.Cells["Id"].Value);
+                    conn.Open();
+                    string query = "DELETE FROM Student WHERE Id = '" + a + "'";
+                    SqlCommand command = new SqlCommand(query, conn);
+                    command.ExecuteNonQuery();
+                    conn.Close();
+                    loadDataGridView();
+                }
             }
             else if (senderGrid.Columns[e.ColumnIndex] is DataGridViewButtonColumn &&
                 e.RowIndex >= 0 && e.ColumnIndex == 6)
@@ -240,7 +246,7 @@ namespace MiniProject
 
         private void button3_Click_1(object sender, EventArgs e)
         {
-            Form_Attendence f = new Form_Attendence();
+            Form_MarkAttendence f = new Form_MarkAttendence();
             f.Show();
             this.Hide();
         }
